### 2022-07-26 #4580 @siavash-saki
@Kevin-Mattheus-Moerman

The new features are introduced in v0.1.0 for doing Clustering analysis. The section "`06 Clustering Urban Areas`" in the documentation is added in v0.1.0 to showcase how one can use tesspy to discretize and cluster city areas. These clusters can represent, for example, functional zones such as residential area, central business district (CBD), etc. However, the final interpretation is open and up to the user. The clustering results can be used as a latent variable in further analyses.

Open Street Map data is used for clustering. Users can choose which data should be used for clustering (by specifying the OSM tags). They can then build a POI count dataframe based on the preferred tiles (tessellation products) (for example, using the `count_poi_per_tile` function) and run the clustering.

In addition, multiple tests are added (and continuously being added), some bugs are fixed, and the documentation is improved.

The changes in the manuscript are described in my comment above.

Thank you!
