### 2022-07-28 #4580 @Kevin-Mattheus-Moerman
Note to handling editor: This is a re-submission of https://github.com/openjournals/joss-reviews/issues/4246, which was initially deemed out of scope. The authors have now expanded the work :point_up: but please ensure reviewers do check if they feel this falls within our scholarly effort criteria.
