### 2022-07-28 #4620 @editorialbot
[REVIEW]: TessPy: a python package for geographical tessellation
**Submitting author:** @siavash-saki (<a href="http://orcid.org/0000-0002-5413-3007">Siavash Saki</a>)
**Repository:** https://github.com/siavash-saki/tesspy
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
