### 2022-07-31 siavash-saki/tesspy#4 @jGaboardi
[JOSS] Rename default branch
xref openjournals/joss-reviews#4620

Rename `master` to `main`. See [the article](https://www.theserverside.com/feature/Why-GitHub-renamed-its-master-branch-to-main) for an explanation of significance.
