### 2022-07-31 siavash-saki/tesspy#5 @jGaboardi
[JOSS] Code formatting in repo?
xref openjournals/joss-reviews#4620

Have you though about using [`black`](https://github.com/psf/black)? It is my personal go-to formatter. It is simple [to use](https://black.readthedocs.io/en/stable/getting_started.html#basic-usage). And can be added with a [pre-commit hook](https://ljvmiranda921.github.io/notebook/2018/06/21/precommits-using-black-and-flake8/) that gets run prior to a PR. This can be run on `.py` files for entire repo in one go, including infrastructure files and notebooks.
