### 2022-07-31 siavash-saki/tesspy#6 @jGaboardi
[JOSS] Test suite failing
xref openjournals/joss-reviews#4620

I ran tests with `pytest` locally twice, each time failed with different output.The test suite fails with the following output:

 <summary>Test run 1</summary>

```
========================================================================= test session starts =========================================================================
platform darwin -- Python 3.10.5, pytest-7.1.2, pluggy-1.0.0 -- /Users/user/miniconda3/envs/py310_tesspy/bin/python
cachedir: .pytest_cache
rootdir: /Users/user/tesspy
plugins: cov-3.0.0
collected 25 items

tests/test_poi_data.py::test_osm_primary_features PASSED                                                                                                        [  4%]
tests/test_poi_data.py::test_create_overpass_query_string PASSED                                                                                                [  8%]
tests/test_poi_data.py::test_get_poi_data PASSED                                                                                                                [ 12%]
tests/test_poi_data.py::test_osm_highway_types PASSED                                                                                                           [ 16%]
tests/test_poi_data.py::test_create_custom_filter PASSED                                                                                                        [ 20%]
tests/test_poi_data.py::test_get_road_network FAILED                                                                                                            [ 24%]
tests/test_tessellation.py::test_squares PASSED                                                                                                                 [ 28%]
tests/test_tessellation.py::test_hexagons PASSED                                                                                                                [ 32%]
tests/test_tessellation.py::test_adaptive_squares PASSED                                                                                                        [ 36%]
tests/test_tessellation.py::test_voronoi FAILED                                                                                                                 [ 40%]
tests/test_tessellation.py::test_city_blocks Fatal Python error: Segmentation fault

Current thread 0x00000001169ff600 (most recent call first):
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/shapely/ops.py", line 160 in unary_union
  File "/Users/user/tesspy/tesspy/tessellation_functions.py", line 422 in get_rest_polygon
  File "/Users/user/tesspy/tesspy/tessellation.py", line 615 in city_blocks
  File "/Users/user/tesspy/tests/test_tessellation.py", line 84 in test_city_blocks
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/python.py", line 192 in pytest_pyfunc_call
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/python.py", line 1761 in runtest
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/runner.py", line 166 in pytest_runtest_call
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/runner.py", line 259 in <lambda>
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/runner.py", line 338 in from_call
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/runner.py", line 258 in call_runtest_hook
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/runner.py", line 219 in call_and_report
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/runner.py", line 130 in runtestprotocol
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/runner.py", line 111 in pytest_runtest_protocol
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/main.py", line 347 in pytest_runtestloop
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/main.py", line 322 in _main
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/main.py", line 268 in wrap_session
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/main.py", line 315 in pytest_cmdline_main
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/config/__init__.py", line 164 in main
  File "/Users/user/miniconda3/envs/py310_tesspy/lib/python3.10/site-packages/_pytest/config/__init__.py", line 187 in console_main
  File "/Users/user/miniconda3/envs/py310_tesspy/bin/pytest", line 11 in <module>

Extension modules: numpy.core._multiarray_umath, numpy.core._multiarray_tests, numpy.linalg._umath_linalg, numpy.fft._pocketfft_internal, numpy.random._common, numpy.random.bit_generator, numpy.random._bounded_integers, numpy.random._mt19937, numpy.random.mtrand, numpy.random._philox, numpy.random._pcg64, numpy.random._sfc64, numpy.random._generator, pandas._libs.tslibs.dtypes, pandas._libs.tslibs.base, pandas._libs.tslibs.np_datetime, pandas._libs.tslibs.nattype, pandas._libs.tslibs.timezones, pandas._libs.tslibs.ccalendar, pandas._libs.tslibs.tzconversion, pandas._libs.tslibs.strptime, pandas._libs.tslibs.fields, pandas._libs.tslibs.timedeltas, pandas._libs.tslibs.timestamps, pandas._libs.properties, pandas._libs.tslibs.offsets, pandas._libs.tslibs.parsing, pandas._libs.tslibs.conversion, pandas._libs.tslibs.period, pandas._libs.tslibs.vectorized, pandas._libs.ops_dispatch, pandas._libs.missing, pandas._libs.hashtable, pandas._libs.algos, pandas._libs.interval, pandas._libs.tslib, pandas._libs.lib, pandas._libs.hashing, pandas._libs.ops, pandas._libs.arrays, pandas._libs.index, pandas._libs.join, pandas._libs.sparse, pandas._libs.reduction, pandas._libs.indexing, pandas._libs.internals, pandas._libs.writers, pandas._libs.window.aggregations, pandas._libs.window.indexers, pandas._libs.reshape, pandas._libs.groupby, pandas._libs.testing, pandas._libs.parsers, pandas._libs.json, pyproj._compat, pyproj._datadir, pyproj._network, pyproj._geod, pyproj.list, pyproj._crs, pyproj._transformer, pyproj.database, pyproj._sync, pygeos.lib, pygeos._geos, pygeos._geometry, shapely.speedups._speedups, sklearn.__check_build._check_build, scipy._lib._ccallback_c, scipy.sparse._sparsetools, scipy.sparse._csparsetools, scipy.sparse.linalg._isolve._iterative, scipy.linalg._fblas, scipy.linalg._flapack, scipy.linalg._cythonized_array_utils, scipy.linalg._flinalg, scipy.linalg._solve_toeplitz, scipy.linalg._matfuncs_sqrtm_triu, scipy.linalg.cython_lapack, scipy.linalg.cython_blas, scipy.linalg._matfuncs_expm, scipy.linalg._decomp_update, scipy.sparse.linalg._dsolve._superlu, scipy.sparse.linalg._eigen.arpack._arpack, scipy.sparse.csgraph._tools, scipy.sparse.csgraph._shortest_path, scipy.sparse.csgraph._traversal, scipy.sparse.csgraph._min_spanning_tree, scipy.sparse.csgraph._flow, scipy.sparse.csgraph._matching, scipy.sparse.csgraph._reordering, sklearn.utils.murmurhash, psutil._psutil_osx, psutil._psutil_posix, numpy.linalg.lapack_lite, scipy.spatial._ckdtree, scipy._lib.messagestream, scipy.spatial._qhull, scipy.spatial._voronoi, scipy.spatial._distance_wrap, scipy.spatial._hausdorff, scipy.special._ufuncs_cxx, scipy.special._ufuncs, scipy.special._specfun, scipy.special._comb, scipy.special._ellip_harm_2, scipy.spatial.transform._rotation, scipy.ndimage._nd_image, _ni_label, scipy.ndimage._ni_label, scipy.optimize._minpack2, scipy.optimize._group_columns, scipy.optimize._trlib._trlib, scipy.optimize._lbfgsb, _moduleTNC, scipy.optimize._moduleTNC, scipy.optimize._cobyla, scipy.optimize._slsqp, scipy.optimize._minpack, scipy.optimize._lsq.givens_elimination, scipy.optimize._zeros, scipy.optimize.__nnls, scipy.optimize._highs.cython.src._highs_wrapper, scipy.optimize._highs._highs_wrapper, scipy.optimize._highs.cython.src._highs_constants, scipy.optimize._highs._highs_constants, scipy.linalg._interpolative, scipy.optimize._bglu_dense, scipy.optimize._lsap, scipy.optimize._direct, scipy.integrate._odepack, scipy.integrate._quadpack, scipy.integrate._vode, scipy.integrate._dop, scipy.integrate._lsoda, scipy.special.cython_special, scipy.stats._stats, beta_ufunc, scipy.stats._boost.beta_ufunc, binom_ufunc, scipy.stats._boost.binom_ufunc, nbinom_ufunc, scipy.stats._boost.nbinom_ufunc, hypergeom_ufunc, scipy.stats._boost.hypergeom_ufunc, ncf_ufunc, scipy.stats._boost.ncf_ufunc, scipy.interpolate._fitpack, scipy.interpolate.dfitpack, scipy.interpolate._bspl, scipy.interpolate._ppoly, scipy.interpolate.interpnd, scipy.interpolate._rbfinterp_pythran, scipy.stats._biasedurn, scipy.stats._levy_stable.levyst, scipy._lib._uarray._uarray, scipy.stats._hypotests_pythran, scipy.stats._statlib, scipy.stats._mvn, scipy.stats._sobol, scipy.stats._qmc_cy, scipy.stats._unuran.unuran_wrapper, sklearn.utils._openmp_helpers, sklearn.utils._logistic_sigmoid, sklearn.utils.sparsefuncs_fast, sklearn.preprocessing._csr_polynomial_expansion, sklearn.utils._typedefs, sklearn.utils._readonly_array_wrapper, sklearn.metrics._dist_metrics, sklearn.metrics.cluster._expected_mutual_info_fast, sklearn.utils._cython_blas, sklearn.utils._heap, sklearn.utils._sorting, sklearn.utils._vector_sentinel, sklearn.metrics._pairwise_distances_reduction, sklearn.metrics._pairwise_fast, sklearn.neighbors._partition_nodes, sklearn.neighbors._ball_tree, sklearn.neighbors._kd_tree, sklearn.decomposition._cdnmf_fast, sklearn.utils._random, sklearn.utils._seq_dataset, sklearn.utils.arrayfuncs, sklearn.linear_model._cd_fast, sklearn._loss._loss, sklearn.utils._weight_vector, sklearn.linear_model._sgd_fast, sklearn.linear_model._sag_fast, sklearn.svm._libsvm, sklearn.svm._liblinear, sklearn.svm._libsvm_sparse, sklearn.decomposition._online_lda_fast, sklearn._isotonic, sklearn.manifold._utils, sklearn.tree._utils, sklearn.tree._tree, sklearn.tree._splitter, sklearn.tree._criterion, sklearn.neighbors._quad_tree, sklearn.manifold._barnes_hut_tsne, sklearn.cluster._k_means_common, sklearn.cluster._k_means_minibatch, sklearn.cluster._k_means_lloyd, sklearn.cluster._k_means_elkan, sklearn.utils._fast_dict, sklearn.cluster._hierarchical_fast, sklearn.cluster._dbscan_inner, hdbscan.dist_metrics, hdbscan._hdbscan_linkage, hdbscan._hdbscan_tree, hdbscan._hdbscan_reachability, hdbscan._hdbscan_boruvka, scipy.cluster._vq, scipy.cluster._hierarchy, scipy.cluster._optimal_leaf_ordering, hdbscan._prediction_utils, h3._cy.util, h3._cy.cells, h3._cy.edges, h3._cy.geo, h3._cy.to_multipoly, matplotlib._c_internal_utils, PIL._imaging, matplotlib._path, kiwisolver._cext, matplotlib._image, _cffi_backend, unicodedata2, rasterio._version, rasterio._err, rasterio._filepath, rasterio._env, rasterio._transform, rasterio._base, rasterio.crs, rasterio._features, rasterio._warp, rasterio._io, osgeo._gdal, osgeo._gdalconst, osgeo._ogr, osgeo._osr, markupsafe._speedups (total: 243)
Segmentation fault: 11
```

 <summary>Test run 2</summary>

```
========================================================================= test session starts =========================================================================
platform darwin -- Python 3.10.5, pytest-7.1.2, pluggy-1.0.0 -- /Users/user/miniconda3/envs/py_tesspy/bin/python
cachedir: .pytest_cache
rootdir: /Users/user/tesspy
plugins: cov-3.0.0
collected 25 items

tests/test_poi_data.py::test_osm_primary_features PASSED                                                                                                        [  4%]
tests/test_poi_data.py::test_create_overpass_query_string PASSED                                                                                                [  8%]
tests/test_poi_data.py::test_get_poi_data PASSED                                                                                                                [ 12%]
tests/test_poi_data.py::test_osm_highway_types PASSED                                                                                                           [ 16%]
tests/test_poi_data.py::test_create_custom_filter PASSED                                                                                                        [ 20%]
tests/test_poi_data.py::test_get_road_network FAILED                                                                                                            [ 24%]
tests/test_tessellation.py::test_squares PASSED                                                                                                                 [ 28%]
tests/test_tessellation.py::test_hexagons PASSED                                                                                                                [ 32%]
tests/test_tessellation.py::test_adaptive_squares PASSED                                                                                                        [ 36%]
tests/test_tessellation.py::test_voronoi FAILED                                                                                                                 [ 40%]
tests/test_tessellation.py::test_city_blocks PASSED                                                                                                             [ 44%]
tests/test_tessellation.py::test_new_shapely PASSED                                                                                                             [ 48%]
tests/test_tessellation.py::test_index_methods PASSED                                                                                                           [ 52%]
tests/test_tessellation.py::test_count_lgu_function_short_version FAILED                                                                                        [ 56%]
tests/test_tessellation.py::test_count_lgu_function_different_city_inputs FAILED                                                                                [ 60%]
tests/test_tessellation.py::test_count_lgu_function_invalid_inputs_city PASSED                                                                                  [ 64%]
tests/test_tessellation.py::test_count_lgu_function_invalid_inputs_gdf PASSED                                                                                   [ 68%]
tests/test_tessellation.py::test_count_lgu_function_invalid_inputs_method PASSED                                                                                [ 72%]
tests/test_tessellation.py::test_count_lgu_function_invalid_inputs_poi PASSED                                                                                   [ 76%]
tests/test_tessellation.py::test_count_lgu_function_different_poi FAILED                                                                                        [ 80%]
tests/test_tessellation.py::test_count_lgu_function FAILED                                                                                                      [ 84%]
tests/test_tessellation.py::test_poi_collection FAILED                                                                                                          [ 88%]
tests/test_tessellation_functions.py::test_count_poi FAILED                                                                                                     [ 92%]
tests/test_tessellation_functions.py::test_get_adaptive_squares FAILED                                                                                          [ 96%]
tests/test_tessellation_functions.py::test_split_linestring FAILED                                                                                              [100%]

============================================================================== FAILURES ===============================================================================
________________________________________________________________________ test_get_road_network ________________________________________________________________________

    def test_get_road_network():
        city = Tessellation("Johannesburg").get_polygon()
        split_road_data = RoadData(city, split_roads=True, verbose=True).get_road_network()
        road_data = RoadData(city, split_roads=False, verbose=True).get_road_network()
        assert len(split_road_data) > 0
        assert len(road_data) > 0
>       assert len(split_road_data) > len(road_data)
E       assert 170 > 170
E        +  where 170 = len(                                            osmid     ref  ... lanes maxspeed\nu          v          key                                  ...               \n87962180   88830454   0                  10429208  CR 391  ...   NaN      NaN\n           89202585   0                  10429208  CR 391  ...   NaN      NaN\n           5762605634 0                 608171338     NaN  ...   NaN      NaN\n88830454   88845187   0                  80565315  CR 391  ...   NaN      NaN\n           564390667  0                  10408441     NaN  ...   NaN      NaN\n...                                           ...     ...  ...   ...      ...\n5764378908 5764378910 0                 608431386     NaN  ...   NaN      NaN\n           5764378921 0    [608431390, 608431391]     NaN  ...   NaN      NaN\n5764378922 5764378921 0                 608431391     NaN  ...   NaN      NaN\n           5764378925 0                 608431397     NaN  ...   NaN      NaN\n5764378921 5764378925 0                 608431393     NaN  ...   NaN      NaN\n\n[170 rows x 10 columns])
[…]
