### 2022-07-31 siavash-saki/tesspy#7 @jGaboardi
[JOSS] Expand on Community guidelines
xref openjournals/joss-reviews#4620, siavash-saki/tesspy#4

There is a good start on this in the docs/README, but I would like to see it more cleaned up and fleshed out. For example, the authors have a [*Contributing to tesspy*](https://github.com/siavash-saki/tesspy#contributing-to-tesspy) section, but it states that contribution should be made from the `main` branch. However, there is currently no `main`, only `master`.
