### 2022-07-31 siavash-saki/tesspy#8 @jGaboardi
[JOSS] Documentation consistency
xref openjournals/joss-reviews#4620

Docstrings are present (and pretty good for most part). I think another comb through for consistent formatting would a long way towards a positive user experience.
