### 2022-07-31 #4620 @jGaboardi
@siavash-saki @martinfleis

I have completed an initial pass for `tesspy`.

This package has the potential to make an immediate impact in the scientific & spatial/geographic communities. There are thoughtful and well-paced tutorial notebooks that provide high-level introductions and more detailed examples for the various tessellation methods offered in `tesspy`. I am excited both for the potential of the package in the long run and possibly to use it in the immediate future. (Also, very cool logo).

Below are some of my initials thoughts from the checklist, etc. that I will begin creating issues for and linking back to this comment.

### Checklist topics
- ***General Checks***
  - *Contribution and authorship*
    - Unclear what role Tobias Hagen has played in the code development (no commits).
    - The repo only 2 PRs; Is there a previous version that is private?
- ***Functionality***
  - *Installation*
    - Install proceeds as planned (on conda-forge). However, tests can't be run out of the box. There needs to testing requirements. What do the authors use to run testing? Maybe `pytest`?
  - *Functionality*
    - Tests do not pass locally for me (see below). I need to see all tests passing before digging deeper into functionality.
- ***Documentation***
  - *A statement of need*
    - The authors define the problem being solved in the docs/README but don't identify the target audience.
  - *Functionality documentation*
    - siavash-saki/tesspy#8
  - *Automated tests*
    - There are tests but they are failing for me locally. I installed following the [`conda` instructions](https://tesspy.readthedocs.io/en/latest/index.html#creating-a-new-environment-for-tesspy).
    - See siavash-saki/tesspy#6 for more information.
  - *Community guidelines*?
    - siavash-saki/tesspy#7
- ***Software paper***
  - I will add more info in here in the coming days.
- ***Misc.*** *consider implementing these before making more substantive changes*
  - siavash-saki/tesspy#4
  - siavash-saki/tesspy#5
