### 2022-07-31 siavash-saki/tesspy#9 @jGaboardi
[JOSS] Supported Python versions?
xref #6 openjournals/joss-reviews#4620

https://github.com/siavash-saki/tesspy/blob/d3e662dfafd887f9ba9bc612d73688515a666af8/setup.py#L23-L28

Currently there are no supported Python versions in `setup.py`. Ideally, this should match the Python version that are being tested against in CI. As an example see [`setup.py` in `pysal/momepy`]().
