### 2022-07-31 siavash-saki/tesspy#10 @jGaboardi
[JOSS] Typos, etc. in the tutorial notebooks
xref openjournals/joss-reviews#4620

- Tutorial 2
  - Missing city blocks plots
- Tutorial 3
  - typo "genrators" above cell 19
- Tutorial 4
  - add contextily to dev requirements
- Tutorial 5
  - second paragraph "Moran's *I*"
  - third paragraph `libpysal`
  - Areas section - "boundary"
  - paragraph under cell 6 "varies"
  - next paragraph "calculated"
