### 2022-07-31 siavash-saki/tesspy#11 @jGaboardi
[JOSS] Dependencies
xref openjournals/joss-reviews#4620

There are packages used directly in `tesspy` that are not included in the requirements. Also, There are a handful of packages that are necessary to run the tutorials. While I know they are critical to install `tesspy` itself, the goal may be to provide a one-stop solution. Maybe adding a `requirements_tutorials.txt` or `requirements_plus.txt` files with these.

* Add to requirements:
  * `scikit-learn`
  * `scipy`
* Needed for tutorials
  *  `contextlily`
  *  `esda`
  *  `libpysal`
  *  `matplotlib`
  *  `seaborn`
  *  `statsmodels`
