### 2022-08-01 siavash-saki/tesspy#12 @jGaboardi
[JOSS] Paper edits, etc.
xref openjournals/joss-reviews#4620

#### Paper sections
- Summary
  - Typos
    - line 11: 'implementations of different methods'
    - line 17/18: 'the overpass API'
  - OSMnx citation
    - line 19: I would keep this JOSS citation and add the add the preferred one:
      - **Boeing, G.** 2017. OSMnx: New Methods for Acquiring, Constructing, Analyzing, and Visualizing Complex Street Networks. Computers, Environment and Urban Systems 65, 126-139. doi:10.1016/j.compenvurbsys.2017.05.004
    - xref #13
- Statement of Need
  - Who is the target audience? There is a *Statement of Need* section, which does an acceptable job at defining the problem, but does not really mention a target audience.
  - line 33: Are the variations actually endless? Maybe something along the lines of 'highly varied flexibility' or the like.
- Related Software packages
  - There needs to be much more reference to related software projects for tessellation, both in Python and other languages. Here are some examples after a quick search:
    - Python
        - [`pytess`](https://pypi.org/project/Pytess/)
        - [`libpysal`](https://github.com/pysal/libpysal) (offers voronoi tessellations in the `cg` module)
    - R
      - [bleiglas](https://doi.org/10.21105/joss.03092): An R package for interpolation and visualisation of spatiotemporal data with 3D tessellation
      - [sf.st_make_grid](https://r-spatial.github.io/sf/reference/st_make_grid.html)
    - Julia
      - [VoronoiDelaunay](https://github.com/JuliaGeometry/VoronoiDelaunay.jl)
