### 2022-08-01 siavash-saki/tesspy#13 @jGaboardi
[JOSS] State of the field / References
xref openjournals/joss-reviews#4620 #12

#### State of the field
- All related and referenced software packages with a DOI discussed in the *Related Software packages* section **or** highlighted in tutorials should cited. A non-exhaustive list includes:
  - PySAL stuff
    - At the very least cite the [`esda` package](https://zenodo.org/record/5139815#.Yucbby-B3fY) and [`libpysal` package](https://zenodo.org/record/6326050#.YucbsC-B3fY).
    - If you are feeling generous cite the [original](https://rrs.scholasticahq.com/article/8285-pysal-a-python-library-of-spatial-analytical-methods) and [most recent](https://onlinelibrary.wiley.com/doi/10.1111/gean.12276) papers.
  - `scikit-learn`
  - [`bleiglas`](https://doi.org/10.21105/joss.03092)
- Also regarding the city block tessellation functionality it would be worth reading [(Thakur & Fan, 2021)](https://escholarship.org/content/qt0kd9q103/qt0kd9q103.pdf?t=r0am1f) (specifically Section 2.3) and referencing their paper.
    * **Thakur, G., & Fan, J.** (2021). MapSpace: POI-based Multi-Scale Global Land Use Modeling. UC Santa Barbara: Center for Spatial Studies. http://dx.doi.org/10.25436/E2Z59N Retrieved from https://escholarship.org/uc/item/0kd9q103
