### 2022-08-02 #4620 @siavash-saki
@jGaboardi

Thank you for carefully reviewing our code and manuscript. We are pleased by your enthusiasm for tesspy and thank you for the insightful comments. We could definitely improve tesspy according to the issues that you opened.

I can answer some of your general questions here:

 - **Unclear what role Tobias Hagen has played in the code development (no commits).**

Tobias Hagen is the project manager of the research project. He has advised and supervised the other two authors along the entire process of code development. His contribution to the TessPy package is raising the research question, pointing out the need for code development, supporting in finding conceptual solutions, reviewing and commenting on ideas and tasks, testing the code on real-world data and checking the usefulness of the code in concrete applications.

 - **The repo only 2 PRs; Is there a previous version that is private?**

The initial version was developed for some parts locally. After that, there were a lot of commits directly to the master branch. But we recently followed the best practice to create, review, and merge Prs.

 - **Install proceeds as planned (on conda-forge). However, tests can't be run out of the box. There needs to testing requirements. What do the authors use to run testing? Maybe pytest?**

We have used pytest.

 - **Tests do not pass locally for me (see below). I need to see all tests passing before digging deeper into functionality.**

We will review the tests and make the necessary changes to ensure they all pass.

Regarding **Documentation**, **Software paper**, and **Misc**:
We will work on the issues in the next days and incorporate the points you mentioned.
