### 2022-08-02 a0566ac
Modified Docstrings, closes #8
