+++ tesspy/tessellation_functions.py
@@ -22,14 +22,14 @@ def count_poi(df,
 
     Parameters
     ----------
-    df : GeoDataFrame
+    df : geopandas.GeoDataFrame
         GeoDataFrame containing the tiles (polygons)
-    points : GeoDataFrame
+    points : geopandas.GeoDataFrame
         GeoDataFrame containing the POI
 
     Returns
     --------
-    final_df : GeoDataFrame
+    final_df : geopandas.GeoDataFrame
         GeoDataFrame containing the tiles and the POI count
     """
     # TODO: Change such that this works for every tessellation method ->
@@ -70,14 +70,14 @@ def get_squares_polyfill(gdf,
 
     Parameters
     ----------
-    gdf : GeoDataFrame
+    gdf : geopandas.GeoDataFrame
         GeoDataFrame containing the tiles (polygons)
     zoom_level : int
         Resolution, which controls the square sizes
 
     Returns
     --------
-    gdf : GeoDataFrame
+    gdf : geopandas.GeoDataFrame
         GeoDataFrame containing the squares
     """
     geom_name = gdf.geometry.name
@@ -123,14 +123,14 @@ def get_h3_hexagons(gdf,
 
     Parameters
     ----------
-    gdf : GeoDataFrame
+    gdf : geopandas.GeoDataFrame
         GeoDataFrame containing the tiles (polygons)
     resolution : int
         Resolution, which controls the hexagon sizes
 
     Returns
     --------
-    gdf : GeoDataFrame
+    gdf : geopandas.GeoDataFrame
         GeoDataFrame containing the hexagons
     """
     if type(gdf.geometry.iloc[0]) == Polygon:
@@ -176,18 +176,18 @@ def get_h3_hexagons(gdf,
 def get_adaptive_squares(input_gdf,
                          threshold):
     """
-    Adaptive tessellation. Subdivides all squares of input tessellation where thresholdi is exceeded.
+    Adaptive tessellation. Subdivides all squares of input tessellation where threshold is exceeded.
 
     Parameters
     ----------
-    input_gdf : GeoDataFrame
+    input_gdf : geopandas.GeoDataFrame
         GeoDataFrame containing the tiles (polygons)
     threshold : int
         threshold, which controls the division of squares
 
     Returns
     --------
-    gdf : GeoDataFrame
+    gdf : geopandas.GeoDataFrame
         GeoDataFrame containing the squares
     """
 
@@ -239,7 +239,7 @@ def voronoi_polygons(sp_voronoi_obj,
     Returns
     --------
     gdf : shapely.Polygon
-        Voronoi polygon
+        Voronoi polygons
     """
     centroid = sp_voronoi_obj.points.mean(axis=0)
 
@@ -289,11 +289,13 @@ def split_linestring(df):
 
     Parameters
     ----------
-    df: DataFrame of shapely.LineStrings containing more than two points
+    df: pandas.DataFrame
+        dataframe of shapely.LineStrings containing more than two points
 
     Returns
     --------
-    dataset: DataFrame with shapely.linestring with two points each
+    dataset: pandas.DataFrame
+        dataframe with shapely.linestring with two points each
     """
     linestrings = []
     osmid = []
@@ -320,11 +322,13 @@ def explode(gdf):
 
     Parameters
     ----------
-    gdf: which can have multi-part geometries that will be exploded
+    gdf:  geopandas.GeoDataFrame
+        GeoDataFrame that can have multi-part geometries that will be exploded
 
     Returns
     --------
-    gdf_out: GeoDataFrame with single geometries
+    gdf_out: geopandas.GeoDataFrame
+        GeoDataFrame with single geometries
     """
[…]
