+++ .github/workflows/tests_package.yml
@@ -5,9 +5,9 @@ name: Python package
 
 on:
   push:
-    branches: [ "main" ]
+    branches: [ "main", "tests-CI" ]
   pull_request:
-    branches: [ "main" ]
+    branches: [ "main", "tests-CI" ]
 
 jobs:
   build:
