+++ .github/workflows/tests_package.yml
@@ -12,7 +12,7 @@ on:
 jobs:
   build:
 
-    runs-on: [ubuntu-latest, macos-latest, windows-2019]
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
