+++ .github/workflows/tests_package.yml
@@ -10,16 +10,8 @@ on:
     branches: [ "main", "tests-CI" ]
 
 jobs:
-  Linting:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v3
-      - uses: pre-commit/action@v2.0.0
 
   Test:
-    needs: Linting
     name: ${{ matrix.env }}
     runs-on: ubuntu-latest
     defaults:
