+++ .github/workflows/draft-pdf.yml
@@ -1,3 +1,9 @@
+name: Paper Draft
+
+on:
+  push:
+    branches: ["main"]
+
 on: [push]
 
 jobs:
