+++ .github/workflows/draft-pdf.yml
@@ -4,8 +4,6 @@ on:
   push:
     branches: ["main"]
 
-on: [push]
-
 jobs:
   paper:
     runs-on: ubuntu-latest
