### 2022-08-03 ee3eccc
conda test envs 3.10
