### 2022-08-03 26a71bc
added conda env with py310 for test
