### 2022-08-03 213d845
Only str input allowed when using count_poi_per_tile()
