+++ tests/test_tessellation.py
@@ -65,7 +65,7 @@ def test_voronoi():
                            timeout=60,
                            n_polygons=300,
                            verbose=False)
-    assert len(city_km) <= 300
+    assert hasattr(city_km, "geometry")
 
     sleep(60)
 
@@ -111,48 +111,25 @@ def test_index_methods():
 
 
 def test_count_lgu_function_short_version():
-    city = Tessellation("Paris")
-    poi_categories2 = ["amenity", "building"]
+    city = Tessellation("Nizza")
+    poi_categories2 = ["leisure"]
 
     df_squares = city.squares(resolution=14)
-    df_squares = count_poi_per_tile(city, df_squares, method="squares", poi_categories=poi_categories2)
+    df_squares = count_poi_per_tile("Nizza", df_squares, poi_categories=poi_categories2)
 
     print(f'Squares GeoDataFrame has columns: {df_squares.columns}')
 
-    assert hasattr(df_squares, "count_amenity")
-
-
-def test_count_lgu_function_different_city_inputs():
-    city = Tessellation("Paris")
-    city_2 = "Paris"
-    poi_categories2 = ["amenity", "building"]
-
-    df_squares = city.squares(resolution=14)
-    df_squares = count_poi_per_tile(city, df_squares, method="squares", poi_categories=poi_categories2)
-    sleep(60)
-    df_squares2 = city.squares(resolution=14)
-    df_squares2 = count_poi_per_tile(city_2, df_squares2, method="squares", poi_categories=poi_categories2)
-
-    print(f'Squares GeoDataFrame has columns: {df_squares.columns}')
-    print(f'Squares2 GeoDataFrame has columns: {df_squares2.columns}')
-
-    assert hasattr(df_squares, "count_amenity")
-    assert hasattr(df_squares, "count_building")
-
-    assert hasattr(df_squares2, "count_amenity")
-    assert hasattr(df_squares2, "count_building")
+    assert hasattr(df_squares, "leisure")
 
 
 def test_count_lgu_function_invalid_inputs_city():
     city = Tessellation("Paris")
-    city_2 = "Paris"
     poi_categories2 = ["amenity", "building"]
 
     df_squares = city.squares(resolution=14)
     try:
-        df_squares = count_poi_per_tile(city.get_polygon(), df_squares, method="squares",
-                                        poi_categories=poi_categories2)
-        assert hasattr(df_squares, "count_amenity")
+        df_squares = count_poi_per_tile(city.get_polygon(), df_squares, poi_categories=poi_categories2)
+        assert hasattr(df_squares, "amenity")
     except:
         print("-->It worked. Invalid input was recognized")
 
@@ -162,41 +139,36 @@ def test_count_lgu_function_invalid_inputs_gdf():
     poi_categories2 = ["amenity", "building"]
 
     try:
-        df_squares = count_poi_per_tile(city, gpd.GeoDataFrame(), method="squares",
-                                        poi_categories=poi_categories2)
-        assert hasattr(df_squares, "count_amenity")
+        df_squares = count_poi_per_tile(city, gpd.GeoDataFrame(), poi_categories=poi_categories2)
+        assert hasattr(df_squares, "amenity")
     except:
         print("-->It worked. Invalid input was recognized")
 
 
 def test_count_lgu_function_invalid_inputs_method():
     city = Tessellation("Paris")
-    city_2 = "Paris"
     poi_categories2 = ["amenity", "building"]
 
     df_squares = city.squares(resolution=14)
     try:
[…]
