### 2022-08-03 96414f4
add dependencies for tutorials
