+++ Examples/requirements_tutorials.txt
@@ -1,0 +1,12 @@
+geopandas=0.10
+scipy
+h3-py
+osmnx
+hdbscan
+mercantile
+scikit-learn
+seaborn
+contextlily
+esda
+libpysal
+statsmodels
