+++ Examples/tutiruals_env.yaml
@@ -1,0 +1,18 @@
+name: test
+channels:
+  - conda-forge
+dependencies:
+  - python=3.9
+  - geopandas=0.10
+  - scipy
+  - h3-py
+  - osmnx
+  - hdbscan
+  - mercantile
+  - scikit-learn
+  # tutorials
+  - seaborn
+  - contextlily
+  - esda
+  - libpysal
+  - statsmodels
