### 2022-08-03 15c152c
update tests CI
Tests are run only for main branch
