### 2022-08-03 siavash-saki/tesspy#6 @jGaboardi
After the recent updates I am still seeing seg fault failures locally (macOS) and another failure on [GHA](https://github.com/jGaboardi/tesspy/runs/7657923410?check_suite_focus=true). This highlights the need for multi-platform testing.
