### 2022-08-04 b9f0b94
changed Test: "test_count_poi". The intial city_squares must be created using get_squares_polyfill because the count_poi function is coming to be tested and this function needs the children id's which are dropen in the standard squares methods
