### 2022-08-04 siavash-saki/tesspy#6 @JoHamann
Hi,
test_count_poi function was updated and should be merged soon. The runtime is still long, this is due to the data collection. We will think about it and maybe combine some tests and use even smaller areas and less POI data that is collected to reduce the overall runtime. This wont affect the quality of the tests
