### 2022-08-04 siavash-saki/tesspy#6 @siavash-saki
@jGaboardi
Thanks for the hints.
- We'll add other platforms.
- There are still some changes to the tests we need to make.
