### 2022-08-04 efe9f29
change tests to get less data from OSM to prevent runtime error (e.g., smaller cities)
