+++ tests/test_tessellation_functions.py
@@ -1,31 +1,28 @@
 from tesspy.poi_data import *
 from tesspy.tessellation import *
 from tesspy.tessellation_functions import *
+import geopandas as gpd
+
+tess_data = gpd.read_file('leisure_poi_Mitte_Berlin.geojson')
+city = Tessellation("Mitte, Berlin")
+city_polygon = city.get_polygon()
 
 
 def test_count_poi():
-    city = Tessellation("Berlin")
-    city_polygon = city.get_polygon()
-    city_poi = POIdata(city_polygon, poi_categories=["leisure"], timeout=60, verbose=False).get_poi_data()
-    points_geom = city_poi[['center_longitude', 'center_latitude']]\
-        .apply(lambda p: Point(p['center_longitude'], p['center_latitude']), axis=1)
-    tess_data = gpd.GeoDataFrame(geometry=points_geom,
-                                 crs='EPSG:4326')
     city_squares = get_squares_polyfill(city_polygon, 14)
     count1 = count_poi(city_squares, tess_data)
+
     assert hasattr(count1, "count")
+    assert hasattr(count1, "children_id")
     assert len(count1) > 0
+    assert count1[count1['quadkey'] == '12021023322022']['count'].iloc[0] == 61
+    assert count1[count1['quadkey'] == '12021023322023']['count'].iloc[0] == 109
+    assert count1[count1['quadkey'] == '12021023322032']['count'].iloc[0] == 56
 
 
 def test_get_adaptive_squares():
-    city = Tessellation("San Diego")
-    city_poi = POIdata(city.get_polygon(), poi_categories=["leisure"], timeout=60, verbose=False).get_poi_data()
-    points_geom = city_poi[['center_longitude', 'center_latitude']] \
-        .apply(lambda p: Point(p['center_longitude'], p['center_latitude']), axis=1)
-    tess_data = gpd.GeoDataFrame(geometry=points_geom,
-                                 crs='EPSG:4326')
-    # we need to use get_squares_polyfill. This method is used within aqk and provides information like children_id
-    city_squares = get_squares_polyfill(city.get_polygon(),14)
+    # we use get_squares_polyfill. This method is used within aqk and provides information like children_id
+    city_squares = get_squares_polyfill(city_polygon, 14)
     count1 = count_poi(city_squares, tess_data)
 
     assert len(get_adaptive_squares(count1, 100)) > len(city_squares)
@@ -33,7 +30,7 @@ def test_get_adaptive_squares():
 
 
 def test_split_linestring():
-    city = Tessellation("Lille")
-    road_data = RoadData(city.get_polygon(), split_roads=False, verbose=True).get_road_network()
+    road_data = pd.read_pickle('road_data_Lille.pkl')
     split_road_data = split_linestring(road_data)
+
     assert len(split_road_data) > len(road_data)
