### 2022-08-04 3e310fb
changed pause time between tests
