+++ .github/workflows/tests_package.yml
@@ -5,7 +5,7 @@ name: Tests
 
 on:
   push:
-    branches: [ "main"]
+    branches: [ "main", "test_branch"]
   pull_request:
     branches: [ "main"]
 
