### 2022-08-04 3ad8dd4
changed pause time between tests
