+++ tests/envs/310-env.yaml
@@ -10,5 +10,6 @@ dependencies:
   - mercantile
   # testing
   - pytest
+  - fiona
   - pytest-cov
   - codecov
