### 2022-08-04 1ceb44a
add exception to catch RuntimeError and wait
