+++ .github/workflows/tests_package.yml
@@ -20,7 +20,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ubuntu-latest
+        os: [ 'ubuntu-latest' ]
         env:
           - tests/envs/38-env.yaml
           - tests/envs/39-env.yaml
