+++ .github/workflows/tests_package.yml
@@ -12,7 +12,7 @@ on:
 jobs:
 
   Test:
-    name: On ${{ matrix.os }} with ${{ matrix.py }}
+    name: Build on ${{ matrix.os }} with ${{ matrix.py }}
     runs-on: ${{ matrix.os }}
     defaults:
       run:
@@ -20,16 +20,19 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ 'ubuntu-latest' ]
-        env:
-          - tests/envs/38-env.yaml
-          - tests/envs/39-env.yaml
-          - tests/envs/310-env.yaml
-        py:
-          - py3.8
-          - py3.9
-          - py3.10
         include:
+          - os: ubuntu-latest
+            env: tests/envs/38-env.yaml
+            py: py3.8
+
+          - os: ubuntu-latest
+            env: tests/envs/39-env.yaml
+            py: py3.9
+
+          - os: ubuntu-latest
+            env: tests/envs/310-env.yaml
+            py: py3.10
+
           - os: macos-latest
             env: tests/envs/310-env.yaml
             py: py3.10
@@ -37,7 +40,6 @@ jobs:
           - os: windows-latest
             env: tests/envs/310-env.yaml
             py: py3.10
-
 
     steps:
       - uses: actions/checkout@v2
