### 2022-08-05 siavash-saki/tesspy#9 @siavash-saki
We have this in the [`setup.py`:](https://github.com/siavash-saki/tesspy/blob/d070497850e623a7968d7c07f7c6353bc8e0bd45/setup.py)

https://github.com/siavash-saki/tesspy/blob/d070497850e623a7968d7c07f7c6353bc8e0bd45/setup.py#L30

I looked at the classifiers in [`momepy/setup.py`:](https://github.com/pysal/momepy/blob/7344e77cda22329a62131ea99a8ffc7ffc6a1b94/setup.py#L36-L42) It's actually the same as ours. I didn't quite understand what to change. Do you mean to add something like these to classifiers?

```
"Programming Language :: Python :: 3.7",
"Programming Language :: Python :: 3.8",
"Programming Language :: Python :: 3.9",
"Programming Language :: Python :: 3.10"
```
