+++ tests/test_tessellation_functions.py
@@ -4,7 +4,7 @@ from tesspy.tessellation_functions import *
 import pandas as pd
 import geopandas as gpd
 
-tess_data = gpd.read_file('tests/leisure_poi_Mitte_Berlin.geojson')
+tess_data = gpd.read_file("tests/leisure_poi_Mitte_Berlin.geojson")
 city = Tessellation("Mitte, Berlin")
 city_polygon = city.get_polygon()
 
@@ -16,9 +16,9 @@ def test_count_poi():
     assert hasattr(count1, "count")
     assert hasattr(count1, "children_id")
     assert len(count1) > 0
-    assert count1[count1['quadkey'] == '12021023322022']['count'].iloc[0] == 61
-    assert count1[count1['quadkey'] == '12021023322023']['count'].iloc[0] == 109
-    assert count1[count1['quadkey'] == '12021023322032']['count'].iloc[0] == 56
+    assert count1[count1["quadkey"] == "12021023322022"]["count"].iloc[0] == 61
+    assert count1[count1["quadkey"] == "12021023322023"]["count"].iloc[0] == 109
+    assert count1[count1["quadkey"] == "12021023322032"]["count"].iloc[0] == 56
 
 
 def test_get_adaptive_squares():
@@ -31,7 +31,7 @@ def test_get_adaptive_squares():
 
 
 def test_split_linestring():
-    road_data = pd.read_pickle('tests/road_data_Lille.pkl')
+    road_data = pd.read_pickle("tests/road_data_Lille.pkl")
     split_road_data = split_linestring(road_data)
 
     assert len(split_road_data) > len(road_data)
