+++ tesspy/poi_data.py
@@ -11,14 +11,16 @@ from shapely.geometry import Point
 
 
 def geom_ceil(coordinate, precision=4):
-    coordinate_ceil = np.true_divide(np.ceil(coordinate * 10 ** precision),
-                                     10 ** precision)
+    coordinate_ceil = np.true_divide(
+        np.ceil(coordinate * 10**precision), 10**precision
+    )
     return coordinate_ceil
 
 
 def geom_floor(coordinate, precision=4):
-    coordinate_floor = np.true_divide(np.floor(coordinate * 10 ** precision),
-                                      10 ** precision)
+    coordinate_floor = np.true_divide(
+        np.floor(coordinate * 10**precision), 10**precision
+    )
     return coordinate_floor
 
 
@@ -41,11 +43,7 @@ class POIdata:
         If True, print information while computing
     """
 
-    def __init__(self,
-                 area,
-                 poi_categories,
-                 timeout,
-                 verbose):
+    def __init__(self, area, poi_categories, timeout, verbose):
         self.area_buffered = None
         self.area = area
         self.poi_categories = poi_categories
@@ -62,35 +60,37 @@ class POIdata:
         --------
         osm_primary_features_lst : list
         """
-        osm_primary_features_lst = ['aerialway',
-                                    'aeroway',
-                                    'amenity',
-                                    'barrier',
-                                    'boundary',
-                                    'building',
-                                    'craft',
-                                    'emergency',
-                                    'geological',
-                                    'healthcare',
-                                    'highway',
-                                    'historic',
-                                    'landuse',
-                                    'leisure',
-                                    'man_made',
-                                    'military',
-                                    'natural',
-                                    'office',
-                                    'place',
-                                    'power',
-                                    'public_transport',
-                                    'railway',
-                                    'route',
-                                    'shop',
-                                    'sport',
-                                    'telecom',
-                                    'tourism',
-                                    'water',
-                                    'waterway']
+        osm_primary_features_lst = [
+            "aerialway",
+            "aeroway",
+            "amenity",
+            "barrier",
+            "boundary",
+            "building",
+            "craft",
+            "emergency",
+            "geological",
+            "healthcare",
+            "highway",
+            "historic",
+            "landuse",
+            "leisure",
+            "man_made",
+            "military",
+            "natural",
+            "office",
+            "place",
+            "power",
+            "public_transport",
+            "railway",
+            "route",
+            "shop",
+            "sport",
+            "telecom",
+            "tourism",
+            "water",
+            "waterway",
+        ]
 
         return osm_primary_features_lst
 
@@ -104,13 +104,10 @@ class POIdata:
         """
         # make the query area a bit larger
         with warnings.catch_warnings():
-            warnings.simplefilter('ignore')
-            self.area_buffered = self.area.buffer(0.008)\
-                                     .simplify(0.005)
+            warnings.simplefilter("ignore")
+            self.area_buffered = self.area.buffer(0.008).simplify(0.005)
 
-        exter_cooridinates = self.area_buffered\
-                                 .iloc[0]\
-                                 .exterior.coords
+        exter_cooridinates = self.area_buffered.iloc[0].exterior.coords
 
         xy = np.array(exter_cooridinates)
 
@@ -131,24 +128,28 @@ class POIdata:
         # todo: is this necessary?
         for poi_category in self.poi_categories:
             if poi_category not in self.osm_primary_features():
[…]
