+++ .github/workflows/tests_package.yml
@@ -11,6 +11,13 @@ on:
 
 jobs:
 
+  Lint:
+    name: Lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: psf/black@stable
+
   Test:
     name: Build on ${{ matrix.os }} with ${{ matrix.py }}
     runs-on: ${{ matrix.os }}
