### 2022-08-05 a590b0a
expand community guideline
