+++ README.md
@@ -99,14 +99,8 @@ All kind of contributions are welcome:
 * Improvement of documentation
 * Additional tests
 
-If you want to contribute to code:
-1. Fork the latest main branch
-2. Create a dev environment: install dependencies and install tesspy in [develop mode](https://python-packaging-tutorial.readthedocs.io/en/latest/setup_py.html#develop-mode)
-3. Write failing tests
-4. Write new code
-5. Run tests and make sure they pass
-6. Update documentation
-7. Pull Request
+Follow the instructions [here](https://tesspy.readthedocs.io/en/latest/Contribution.html)
+for submitting a PR.
 
 If you have any ideas or questions, feel free to open an issue.
 
