+++ setup.py
@@ -25,7 +25,10 @@ setuptools.setup(
     url="https://github.com/siavash-saki/tesspy",
     packages=setuptools.find_packages(),
     classifiers=[
-        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
         "License :: OSI Approved :: BSD License",
         "Intended Audience :: Science/Research",
         "Operating System :: OS Independent",
