### 2022-08-05 870aefb
add 7th example notebook
