+++ docs/source/Examples.rst
@@ -11,5 +11,6 @@ Examples
    04 City Blocks <example_notebooks/04_City_Blocks>
    05 Analyze Urban Areas <example_notebooks/05_Analyze_Urban_Areas>
    06 Clustering Urban Areas <example_notebooks/06_Clustering_Urban_Areas>
+   07 POI counts <example_notebooks/07_LocalGeographicUnits>
 
 
