### 2022-08-08 siavash-saki/tesspy#30 @jGaboardi
[JOSS] import error in 01_Getting_Started
xref https://github.com/openjournals/joss-reviews/issues/4620

The first code cell in [01_Getting_Started](https://tesspy.readthedocs.io/en/latest/example_notebooks/01_Getting_Started.html#Tessellation-of-Frankfurt-am-Main) fails due to the following, which should be importing from `tesspy`:

```python
from tessellation import Tessellation
```

Both the notebook in [Examples](https://github.com/siavash-saki/tesspy/blob/main/Examples/01_Getting_Started.ipynb) and the [docs](https://github.com/siavash-saki/tesspy/blob/main/docs/source/example_notebooks/01_Getting_Started.ipynb) should be updated.
