### 2022-08-11 #4620 @BenjMy
@siavash-saki @martinfleis

I have completed an initial pass for tesspy (using the dev environment, on ubuntu 20.04).

Coming from the Geosciences community, I didn't know about Tesselation and appreciated the opportunity to read and review this paper and the code. I found the paper a little bit technical but it reads quite well (see my minor comments below). As for the code, I found some minor issues also listed below.
As the authors state in the manuscript that `tesspy` may be used to "find the most suitable method" my main concern for both code/article is the lack of metrics/explanation on how to evaluate the quality of the Tesselation results.

I thank in advance the authors for their response and hope to see the manuscript published in JOSS soon.

### Checklist topics

- #### Functionality
  - _Installation_
  As the Conda package is not up to date I had to go through a [dev environment ](https://tesspy.readthedocs.io/en/latest/Contribution.html#create-a-dev-env) in order to be able to run the examples:

    - How do you plan to release on Conda? If not using CI, I recommend adding instructions on how to install the development version using setup.py in the README directly (not only in the documentation)
    - Also there is a need to add instructions on **how to install notebook examples** dependencies and **how to run them** using jupyter-notebooks. This was not very intuitive to me in particular because there are both yaml and requirements.txt files in the example folder.
    - Notebooks installation returns errors due to the wrong syntax in requirement.txt (check this also for the tutorials_env.yaml
):
      - ERROR: Invalid requirement: 'geopandas=0.10' (from line 1 of requirements_tutorials.txt). Hint: = is not a valid operator. Did you mean == ?
      - Replace h3-py by h3 (?)

  - _Functionality_
    - No issue during the execution of the notebooks except for `01_Getting_Started` due to the already raised issue https://github.com/siavash-saki/tesspy/issues/30#issue-1331166797
    - I successfully tested `tesspy` for my hometown city. It was really interesting and easy to get results. I appreciated that data are automatically fetched using API from OpenStreetMap.

- #### Documentation
  - _Functionality documentation_
  As the authors state in the manuscript that `tesspy` may be used to "find the most suitable method", I wonder if the code includes a way to evaluate this. Is there existing metrics for that purpose? as it is done for meshing (i.e. aspect ratio, skewness, orthogonality, and smoothness).
  - _Automated tests_
I run the tests locally, and had the same report than the authors using their Github action tests_package.yml i.e. 27 passed and 22 warnings (not attributed to `tesspy`) with a significant coverage.

- #### Software paper
  - I have a couple of questions/ minor issues regarding the paper.
    - L.44: Is there a proxy to evaluate the quality of the different tesselation (or unit) with respect to a given topology?
    - Fig. 3 right: I don't understand what the 'spatial lag' analysis consists in. Can you say a word about it in the text? is it the result spatial autocorrelation of fig3 left?
    - L.59/60: You state "to easily find the most suitable method" based on what? see my comment for L.44.

  - _State of the field_
    - How do you position `tesspy`compare to this [package](https://nurbs-python.readthedocs.io/en/latest/module_tessellate.html#:~:text=The%20tessellation%20function%20is%20designed,generated%20from%20the%20input%20vertices.)
    - Should you also include references to other Open-Source packages such as [Tess in R](https://search.r-project.org/CRAN/refmans/spatstat.geom/html/tess.html)  (?)
