### 2022-08-11 siavash-saki/tesspy#37 @BenjMy
Notebooks installation returns errors due to the wrong syntax in requirement.txt
xref https://github.com/openjournals/joss-reviews/issues/4620

During the installation of the [requirements_tutorials.txt](https://github.com/siavash-saki/tesspy/blob/main/Examples/requirements_tutorials.txt) file using `pip install -r requirements_tutorials.txt`:

- an error is raised due to the wrong operator specifying geopandas version
- h3-py is not existing (I assume this should be replaced by h3)

You should also check this also for the tutorials_env.yaml
By the way there is a typo in the filename, replace tutiruals_env.yaml by tutorials_env
