### 2022-08-11 siavash-saki/tesspy#37 @siavash-saki
The dependency is called "h3" on pypi and "h3-py" in conda-forge. The `requirements_tutorials.txt` is supposed to be used for installing the dependencies using conda.

You can actually create a new conda env with channel set to `conda-forge` and install the requirements using conda as follows:

`$ conda install -c conda-forge --file requirements_tutorials.txt`

That should actually work. (I just tested it again). And the operator specifying the geopandas version doesn't raise any error in this method.

I will add this instruction also the the example notebook.

Btw, thanks for mentioning the typo in the file name. I'll correct it.
