### 2022-08-12 siavash-saki/tesspy#41 @siavash-saki
[JOSS] adding instructions on how to install the dev version in the README
xref https://github.com/openjournals/joss-reviews/issues/4620 https://github.com/siavash-saki/tesspy/issues/12

> I recommend adding instructions on how to install the development version using setup.py in the README directly (not only in the documentation)
