### 2022-08-12 siavash-saki/tesspy#42 @siavash-saki
[JOSS] State of the field / consider these packages
xref https://github.com/openjournals/joss-reviews/issues/4620 https://github.com/siavash-saki/tesspy/issues/12

> - How do you position tesspy compare to this [package](https://nurbs-python.readthedocs.io/en/latest/module_tessellate.html#:~:text=The%20tessellation%20function%20is%20designed,generated%20from%20the%20input%20vertices.)
> - Should you also include references to other Open-Source packages such as [Tess in R](https://search.r-project.org/CRAN/refmans/spatstat.geom/html/tess.html) (?)
