### 2022-08-12 siavash-saki/tesspy#43 @siavash-saki
[JOSS] software paper - fig3
xref https://github.com/openjournals/joss-reviews/issues/4620 https://github.com/siavash-saki/tesspy/issues/12

> Fig. 3 right: I don't understand what the 'spatial lag' analysis consists in. Can you say a word about it in the text? is it the result spatial autocorrelation of fig3 left?
