### 2022-08-12 #4620 @siavash-saki
@BenjMy

Thank you very much for putting time and effort into reviewing `tesspy`. Each and every comment of yours is genuinely improving this package, as we've seen during the past few days.

I created issues for the tasks you mentioned. They will be addressed in the next few days. I can also answer some of your general questions here:

- Installation:
  - We currently use CI to release on PyPI but manually publish on conda. The v0.1.1 came just a few hours ago on conda-forge. But it would be a good idea to also consider CI for releasing on conda. Anyway, an instruction about setting up a dev version on Readme is a good idea. I'll do it.
  - An instruction is added to the "01_Getting_Started" notebook, which explains how to install `requirements` and how to run example notebooks.
  - The confusion about the syntax error is also explained and added to the "01_Getting_Started" notebook.
- Functionality:
  - Import error is fixed.
- Paper:
  - About your question: Based on what we can find the best tessellation method? This is a very complex and interesting research question. Unfortunately, there is not a short answer to this. There isn't any global metric for all use cases to demonstrate the best tessellation. We are actually publishing a paper with the same research question in a few weeks. In that paper, we choose a specific use case and dive deep into comparing different tessellation methods and their effects on the spatial analyses. For example, we show how different methods alter the underlying distribution of features and how it could affect our final results. However, I'm afraid that would be too technical in the context of `tesspy`. `Tesspy` aims to provide a simple and easy-to-use spatial framework to explore different tessellation variations and discover what works bests using experiments or try-and-error. In addition, the more advanced users can do more specific statistical analyses based on their dataset and use cases to find the proper tessellation method.
  - I'll add a few words about the fig3 and spatial lag.
