+++ paper/paper.bib
@@ -9,6 +9,15 @@
   publisher={Elsevier}
 }
 
+@misc{bahgat_2022,
+ author = {Bahgat, Karim and Currie, Michael},
+ year = {2022},
+ note = {Accessed: 2022-07-01},
+ title = {Pytess: Pure Python tessellation of points into polygons, including Delauney/Thiessin, and Voronoi polygons},
+ url = {https://github.com/karimbahgat/Pytess},
+ journal = {GitHub}
+}
+
 @article{belej:2020,
   title={3D Modeling of Discontinuity in the spatial distribution of apartment prices using voronoi diagrams},
   author={Be{\l}ej, Miros{\l}aw and Figurska, Marta},
@@ -21,6 +30,17 @@
   publisher={Multidisciplinary Digital Publishing Institute}
 }
 
+@article{bingol2019geomdl,
+  title={{NURBS-Python}: An open-source object-oriented {NURBS} modeling framework in {Python}},
+  author={Bingol, Onur Rauf and Krishnamurthy, Adarsh},
+  journal={{SoftwareX}},
+  volume={9},
+  pages={85--94},
+  year={2019},
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.softx.2018.12.005}
+}
+
 @article{boeing:2017,
   title={OSMnx: A Python package to work with graph-theoretic OpenStreetMap street networks},
   author={Boeing, Geoff},
@@ -29,6 +49,46 @@
   number={12},
   doi={10.21105/joss.00215},
   year={2017}
+}
+
+@article{boeing_2017_2,
+  title={OSMnx: New methods for acquiring, constructing, analyzing, and visualizing complex street networks},
+  author={Boeing, Geoff},
+  journal={Computers, Environment and Urban Systems},
+  volume={65},
+  pages={126--139},
+  year={2017},
+  doi={10.1016/j.compenvurbsys.2017.05.004},
+  publisher={Elsevier}
+}
+
+@software{esda_2021,
+title={esda: Exploratory Spatial Data Analysis in PySAL},
+url={https://github.com/pysal/esda},
+year = 2021,
+publisher = {Zenodo},
+version = {v2.4.1},
+doi = {10.5281/zenodo.5139815},
+author = {Sergio Rey and
+	Levi John Wolf and
+	Dani Arribas-Bel and
+	James Gaboardi and
+	Wei Kang and
+	mhwang4 and
+	jlaura and
+	pastephens and
+	Charles Schmidt and
+	Stefanie Lumnitz and
+	David C. Folch and
+	Martin Fleischmann and
+	Juan C Duque and
+	Luc Anselin and
+	Nicholas Malizia and
+	eli knaap and
+	Filipe and
+	matthewborish and
+	Leo Morales and
+	Mridul Seth}
 }
 
 @article{fleischmann:2019,
@@ -42,12 +102,32 @@
     DOI={10.21105/joss.01807}
 }
 
+@misc{h3_uber,
+title={h3-py: Uber's H3 Hexagonal Hierarchical Geospatial Indexing System in Python},
+url={https://github.com/uber/h3-py},
+journal={GitHub},
+note = {Accessed: 2022-07-01},
+year = {2022},
+author={{h3-py Contributors}}
+}
+
 @article{hagen:2022,
   title={Discretization of Urban Areas using POI-based Tessellation},
   author={Hagen, Tobias and Hamann, Jonas and Saki, Siavash},
   year={2022},
   publisher={Frankfurt University of Applied Sciences},
   DOI={10.48718/7jjr-1c66}
+}
+
+@article{hdbscan_2017,
+  title={hdbscan: Hierarchical density based clustering},
[…]
