+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 Discretization of urban areas is a crucial aspect in many spatial analyses. It helps to understand geographical space and provides a framework for analyzing geospatial data [@gold:2016]. The process of discretization of space into subspaces without overlaps and gaps is called tessellation. The polygons, created using tessellation, are called tiles. Tiles can have the same shape and size (regular) or may differ in shape or size (irregular).
 
-`TessPy` contains implementations of different tessellation methods for geographical areas and is designed to be flexible and easy to use. It is built on top of `GeoPandas` [@kelsey_jordahl_2019]. Tiles are returned in a `GeoDataFrame` in `shapely` [@shapely_software] `Polygon` format. Regular methods implemented here are squares (based on @Mercantile) and hexagons (based on @h3_uber). The implemented irregular tessellations are adaptive squares, Voronoi diagrams, and city blocks. The irregular methods are data-driven and the required data is retrieved from the OpenStreetMap (OSM) [@OpenStreetMap]. Points of Interest (POI) are downloaded using the `overpass` API, the official OSM API, to access the OSM database. To get the road network data, `osmnx` [@boeing:2017] is used (more info at @boeing_2017_2). The input can be in different formats, e.g., a `Polygon` (or a `Multipolygon`) with a defined coordinate reference system (CRS), or an address.  More methods and functions are explained in the API documentation. Several examples (e.g., clustering of urban areas) are demonstrated in `TessPy` documentation. Clustering algorithms within the tutorials and the methods use scikit-learn [@scikit-learn] and hdbscan [@hdbscan_2017]. Further packages for conducting spatial analyses in the tutorials are @esda_2021 and @libpysal_2021, which are subsets of @pysal_2022 (More info to be found at @rey2010pysal and @rey2021pysal).
+`TessPy` contains implementations of different tessellation methods for geographical areas and is designed to be flexible and easy to use. It is built on top of `GeoPandas` [@kelsey_jordahl_2019]. Tiles are returned in a `GeoDataFrame` in `shapely` [@shapely_software] `Polygon` format. Regular methods implemented here are squares (based on ` Mercantile` [@Mercantile]) and hexagons (based on `h3-py` [@h3_uber]). The implemented irregular tessellations are adaptive squares, Voronoi diagrams, and city blocks. The irregular methods are data-driven and the required data is retrieved from the OpenStreetMap (OSM) [@OpenStreetMap]. Points of Interest (POI) are downloaded using the `overpass` API, the official OSM API, to access the OSM database. To get the road network data, `osmnx` [@boeing:2017] is used (more info at @boeing_2017_2). The input can be in different formats, e.g., a `Polygon` (or a `Multipolygon`) with a defined coordinate reference system (CRS), or an address.  More methods and functions are explained in the API documentation. Several examples (e.g., clustering of urban areas) are demonstrated in `TessPy` documentation. Clustering algorithms within the tutorials and the methods use scikit-learn [@scikit-learn] and hdbscan [@hdbscan_2017]. Further packages for conducting spatial analyses in the tutorials are `esda` [@esda_2021] and `libpysal`  [@libpysal_2021], which are subsets of `pysal` [@pysal_2022] (More info to be found at @rey2010pysal and @rey2021pysal).
 
 Being extensive and flexible, `TessPy` allows the user to customize the tessellation methods. This is essential, especially when it comes to irregular methods. For example, different initial clustering algorithms can be used to define centroids in the Voronoi method. \autoref{fig1} shows three examples for irregular tessellations for Frankfurt am Main.
 
