+++ paper/paper.bib
@@ -370,7 +370,7 @@ author={Want, Chris}
 }
 
 @misc{VoronoiDelaunay_jl_2022,
- author = {VoronoiDelaunay.jl Contributors},
+ author = {{VoronoiDelaunay.jl Contributors}},
  year = {2022},
  note = {Accessed: 2022-07-01},
  title = {VoronoiDelaunay.jl: Fast and robust Voronoi & Delaunay tessellation creation with Julia},
