### 2022-08-17 74ae610
pysal bib title fixed
