+++ paper/paper.bib
@@ -239,7 +239,7 @@ author={{Mercantile Contributors}}
 }
 
 @software{pysal_2022,
-title={libpysal: Python Spatial Analysis Library Core},
+title={PySAL: Python Spatial Analysis Library Meta-Package },
 url={https://github.com/pysal/pysal},
 year = 2022,
 publisher = {Zenodo},
