### 2022-08-17 #4620 @siavash-saki
@jGaboardi
Thanks for the info. We actually addressed all the JOSS-related issues.
I'd appreciate it if you could look over the revised paper.
