+++ paper/paper.bib
@@ -38,7 +38,7 @@
   pages={85--94},
   year={2019},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.softx.2018.12.005}
+  doi={10.1016/j.softx.2018.12.005}
 }
 
 @article{boeing:2017,
