### 2022-08-17 c1c57fb
update shapely citation
