+++ paper/paper.bib
@@ -331,13 +331,11 @@ author = {Sergio Rey and
   number = {1},
 }
 
-@misc{shapely_software,
-title={Shapely/shapely: Manipulation and analysis of geometric objects},
-url={https://github.com/shapely/shapely},
-journal={GitHub},
-note = {Accessed: 2022-07-01},
-year = {2022},
-author={{Shapely Contributors}}
+@Misc{shapely,
+  author =    {Sean Gillies and others},
+  title =     {Shapely: manipulation and analysis of geometric objects},
+  year =      {2007--},
+  url =       {https://github.com/shapely/shapely},
 }
 
 @article{spatstat_package_2005,
