### 2022-08-18 #4620 @BenjMy
@siavash-saki Thanks for your answers to my questions.

@martinfleis I've read the revised manuscript and I'm satisfied with the changes. All the criteria from the checklist are met. For me it is ready for publication.

All the best to `TessPy` and the dev team!
