+++ setup.py
@@ -16,7 +16,7 @@ INSTALL_REQUIRES = [
 
 setuptools.setup(
     name="tesspy",
-    version="0.1.1",
+    version="0.1.2",
     author="tesspy Developers",
     author_email="jonas.hamann@fb3.fra-uas.de",
     description="Tessellation of Urban Areas",
