### 2022-08-23 #4620 @siavash-saki
@martinfleis

- [x] A new tagged release [v0.1.2](https://github.com/siavash-saki/tesspy/releases/tag/v0.1.2) with the latest changes is created.
- [x] The reviewed software (latest release) is [archived in Zenodo](https://zenodo.org/record/7016503#.YwSda2FByB8).  [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7016503.svg)](https://doi.org/10.5281/zenodo.7016503)
- [x] Metadata is modified and checked. The title and author list match the metadata of the paper.
- [x] DOI of the archived version in Zenodo: `10.5281/zenodo.7016503`
