### 2022-08-26 #4620 @siavash-saki
@jGaboardi @BenjMy @martinfleis
I sincerely appreciate the time and effort you put into reviewing the submissions (especially our submission). Publishing in JOSS was a unique experience, and the review process was extremely smooth and helpful.
We would be more than happy if we could ever contribute to JOSS. (Donation is already made) :D
I wish you all the best.
